Fix delete address and update info

+++ server/public/js/profile.js
@@ -3,7 +3,7 @@ const editUser = async (event) => {
   const userInfo = getAllFormFields();
   const isValid = checkIfValid(userInfo);
   if (isValid) {
-    const response = await fetch(`/api/user/${id}`, {
+    const response = await fetch(`/api/user/${userInfo.id}`, {
       method: "PUT",
       body: JSON.stringify(userInfo),
       headers: { "Content-Type": "application/json" },
@@ -14,50 +14,55 @@ const editUser = async (event) => {
     } else {
       alert(response.statusText);
     }
-  } {
-    alert(`Please check out invalid/empty fields: ${invalidFields.join(",")}`);
-    invalidFields = [];
   }
-}
-
-const getAllFormFields = () => {
-  const name = document.querySelector("#name-signup").value.trim();
-  const email = document.querySelector("#email-signup").value.trim();
-  const password = document.querySelector("#password-signup").value.trim();
-  return { name, email, password }
-}
-
-
-
-  const deleteAddress = async (event) => {
-    event.preventDefault();
-    console.log(1)
-    const id = event.target.getAttribute("data-id");
-    console.log(id)
-    // Send a DELETE request to the API endpoint
-    const response = await fetch(`/api/address/${id}`, {
-        method: "DELETE",
-        headers: { "Content-Type": "application/json" },
-    });
-
-    if (response.ok) {
-        // If successful, redirect the browser to the profile page
-        document.location.replace("/profile");
-    } else {
-        alert(response.statusText);
-    }
-
 };
 
-const getShipmentStatus = () => {
+const getAllFormFields = () => {
+  const name = document.querySelector("#user-name").value.trim();
+  const email = document.querySelector("#email").value.trim();
+  const id = document.querySelector("#edit-btn").dataset.id;
+  return { name, email, id };
+};
+const invalidFields = [];
+const checkIfValid = (addressInfo) => {
+  for (fieldName in addressInfo) {
+    if (addressInfo[fieldName] == "" || !addressInfo[fieldName]) {
+      invalidFields.push(fieldName);
+    }
+  }
+  if (invalidFields.length > 0) {
+    return false;
+  }
+  return true;
+};
 
+const deleteAddress = async (event) => {
+  event.preventDefault();
+  const id = event.currentTarget.getAttribute("data-id");
+  console.log(id);
+  if (id === null) {
+    alert("null id ");
+    return;
+  }
+  // Send a DELETE request to the API endpoint
+  const response = await fetch(`/api/address/${id}`, {
+    method: "DELETE",
+    headers: { "Content-Type": "application/json" },
+  });
+
+  if (response.ok) {
+    // If successful, redirect the browser to the profile page
+    document.location.replace("/profile");
+  } else {
+    alert(response.statusText);
+  }
+};
+
+const getShipmentStatus = () => {};
+
+document.getElementById("edit-btn").addEventListener("click", editUser);
+
+const deleteEl = document.querySelector(".address-delete");
+if (deleteEl) {
+  deleteEl.addEventListener("click", deleteAddress);
 }
-
-
-// document
-//   .getElementById("edit-btn")
-//   .addEventListener("click", editUser);
-
-document
-    .querySelector(".address-delete")
-    .addEventListener("click", deleteAddress);
